requirements.txt
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
Django==3.2.5
django-cors-headers==3.7.0
django-modeltranslation==0.17.3
djangorestframework==3.12.4
djangorestframework-simplejwt
django-extensions==2.2.5
google-api-core==1.31.1
google-auth==1.34.0
google-cloud-secret-manager==2.6.0
googleapis-common-protos==1.53.0
gql==3.0.0a6
graphql-core==3.1.5
grpc-google-iam-v1==0.12.3
grpcio==1.39.0
idna==3.2
libcst==0.3.19
multidict==5.1.0
mypy-extensions==0.4.3
packaging==21.0
Pillow
promise==2.3
proto-plus==1.19.0
protobuf==3.17.3
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyOpenSSL==19.0.0
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
rsa==4.7.2
Rx==1.6.1
six==1.16.0
social-auth-app-django==5.0.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
typing-inspect==0.7.1
urllib3==1.26.6
websockets==9.1
werkzeug==0.16.0
yarl==1.6.3
